+++ tests/test_formfields.py
@@ -1,7 +1,7 @@
 from unittest import mock
 
 from django import forms
-from django.test import SimpleTestCase
+from django.test import SimpleTestCase, override_settings
 from django.utils.functional import lazy
 
 from phonenumber_field.formfields import PhoneNumberField
@@ -18,6 +18,23 @@ class PhoneNumberFormFieldTest(SimpleTestCase):
         self.assertIs(form.is_valid(), False)
         self.assertEqual(
             form.errors, {"number": ["Enter a valid phone number (e.g. +12125552368)."]}
+        )
+
+    @override_settings(PHONENUMBER_DEFAULT_REGION="FR")
+    def test_error_message_uses_default_region(self):
+        class PhoneNumberForm(forms.Form):
+            number = PhoneNumberField()
+
+        form = PhoneNumberForm({"number": "invalid"})
+        self.assertIs(form.is_valid(), False)
+        self.assertEqual(
+            form.errors,
+            {
+                "number": [
+                    "Enter a valid phone number (e.g. 01 23 45 67 89) "
+                    "or a number with an international call prefix."
+                ]
+            },
         )
 
     def test_override_error_message(self):